+++ packages/components/atoms/f-form-field/test/specs/accessibility/axe-accessibility.spec.js
@@ -1,10 +1,15 @@
-import { getAccessibilityTestResults, processResults } from '../../../../../../../test/utils/axe-helper';
+import { getAccessibilityTestResults } from '../../../../../../../test/utils/axe-helper';
+import FormFieldComponent from '../../../test-utils/component-objects/f-form-field.component';
 
 describe('Accessibility tests', () => {
-    it.skip('a11y - should test f-formField component WCAG compliance', () => {
+    beforeEach(() => {
+        browser.url('?path=/story/components-atoms--form-field-component');
+        browser.switchToFrame(0);
+        FormFieldComponent.waitForFormField();
+    });
+
+    it('a11y - should test f-formField component WCAG compliance', () => {
         // Act
         const axeResults = getAccessibilityTestResults('f-form-field');
-        // Assert
-        expect(axeResults.violations.length).toBe(0);
     });
 });