+++ packages/components/molecules/f-alert/jest.config.js
@@ -35,7 +35,8 @@ module.exports = {
     testURL: 'http://localhost/',
 
     modulePathIgnorePatterns: [
-        './test/specs/component/'
+        './test/specs/component/',
+        './test/specs/accessibility'
     ],
 
     setupFilesAfterEnv: [