+++ packages/components/organisms/f-checkout/src/store/checkout.module.js
@@ -56,7 +56,7 @@ export default {
 
             const { data } = await axios.get(url, config);
 
-            commit('updateState', data);
+            commit('UPDATE_STATE', data);
         },
 
         /**
@@ -134,16 +134,24 @@ export default {
 
             const { data } = await axios.get(url, config);
 
-            commit('updateAvailableFulfilment', data);
+            commit('UPDATE_AVAILABLE_FULFILMENT_TIMES', data);
         },
 
         setAuthToken: ({ commit }, authToken) => {
-            commit('updateAuth', authToken);
+            commit('UPDATE_AUTH', authToken);
+        },
+
+        updateFulfilmentAddress ({ commit }, payload) {
+            commit('UPDATE_FULFILMENT_ADDRESS', payload);
+        },
+
+        updateMobileNumber ({ commit }, payload) {
+            commit('UPDATE_MOBILE_NUMBER', payload);
         }
     },
 
     mutations: {
-        updateState: (state, {
+        UPDATE_STATE: (state, {
             id,
             serviceType,
             customer,
@@ -179,7 +187,7 @@ export default {
             state.messages = messages;
         },
 
-        updateAvailableFulfilment: (state, {
+        UPDATE_AVAILABLE_FULFILMENT_TIMES: (state, {
             times,
             asapAvailable
         }) => {
@@ -187,9 +195,20 @@ export default {
             state.availableFulfilment.isAsapAvailable = asapAvailable;
         },
 
-        updateAuth: (state, authToken) => {
+        UPDATE_AUTH: (state, authToken) => {
             state.authToken = authToken;
             state.isLoggedIn = !!authToken;
+        },
+
+        UPDATE_FULFILMENT_ADDRESS (state, address) {
+            state.fulfilment.address = {
+                ...state.fulfilment.address,
+                ...address
+            };
+        },
+
+        UPDATE_MOBILE_NUMBER (state, mobileNumber) {
+            state.customer.mobileNumber = mobileNumber;
         }
     }
 };