+++ packages/tools/generator-component/generators/app/index.js
@@ -48,8 +48,8 @@ module.exports = class extends Generator {
                                 .replace(/(f-skeleton)/g, `f-${nameTransformations.class}`)
                                 .replace(/__/g, ''); // We don't want to have file names such as .test.js or .stories.js, otherwise Jest or Storybook will pick them up from the templates folder.
         }));
-        let ignoreTestPattern = this.answers.needsComponentTests ? [] : ["**/*/test/specs/component", '**/*/test-utils/component-objects']
-        const ignoreApiMockPattern = this.answers.needsTestingApiMocks ? [] : ["**/*/src/services"];
+        let ignoreTestPattern = this.answers.needsComponentTests ? [] : ['**/*/test/specs/component'];
+        const ignoreApiMockPattern = this.answers.needsTestingApiMocks ? [] : ['**/*/src/services'];
     
         ignoreTestPattern = ignoreTestPattern.concat(ignoreApiMockPattern);